+++ index.js
@@ -93,7 +93,13 @@ export { default as SearchField } from './lib/SearchField';
 
 /* specific use */
 export { default as FilterPane } from './lib/FilterPane';
-export { default as FilterGroups, filterState, filters2cql, onChangeFilter } from './lib/FilterGroups';
+export {
+  default as FilterGroups,
+  filterState,
+  filters2cql,
+  onChangeFilter,
+  handleFilterChange
+} from './lib/FilterGroups';
 export { default as FilterControlGroup } from './lib/FilterControlGroup';
 export { default as FilterPaneSearch } from './lib/FilterPaneSearch';
 export { default as ExportCsv } from './lib/ExportCsv';