+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6.1.1 (IN-PROGRESS)
 
 * Fix issue with `initialStatus` prop on Accordions not working.
+* Export `handleFilterChange` at top level.
 
 ## [6.1.0](https://github.com/folio-org/stripes-components/tree/v6.1.0) (2020-03-16)
 [Full Changelog](https://github.com/folio-org/stripes-components/compare/v6.0.0...v6.1.0)